pyproject.toml
[tool.poetry]
name = "fastapi-auth-session"
version = "0.1.0"
description = ""
authors = ["andimeon <andimeon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
fastapi = "^0.109.0"
sqlalchemy = "^2.0.25"
jinja2 = "^3.1.3"
sqladmin = "^0.16.0"
alembic = "^1.13.1"
uvicorn = "^0.27.0"
pydantic-settings = "^2.1.0"
ipython = "^8.20.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pytz = "^2024.1"
sqlalchemy-utils = "^0.41.1"
python-multipart = "^0.0.9"
itsdangerous = "^2.1.2"
pydantic = {extras = ["email"], version = "^2.6.4"}
passlib = "^1.7.4"
bcrypt = "^4.1.2"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
types-requests = "^2.31.0.20240310"
types-redis = "^4.6.0.20240218"
pydocstyle = "^6.3.0"
flake8 = "^7.0.0"
flake8-2020 = "^1.8.1"
flake8-blind-except = "^0.2.1"
flake8-bugbear = "^24.2.6"
flake8-builtins = "^2.2.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.14.0"
flake8-debugger = "^4.1.2"
flake8-docstrings = "^1.7.0"
flake8-eradicate = "^1.5.0"
flake8-polyfill = "^1.0.2"
flake8-print = "^5.0.0"
flake8-quotes = "^3.4.0"
flake8-string-format = "^0.3.0"
flake8-fixme = "^1.1.1"
flake8-annotations-complexity = "^0.0.8"
flake8-variables-names = "^0.0.6"
flake8-class-attributes-order = "^0.1.3"
flake8-broken-line = "^1.0.0"
flake8-tidy-imports = "^4.10.0"
flake8-typing-imports = "^1.15.0"
dlint = "^0.14.1"
flake8-if-statements = "^1.0.0"
flake8-functions = "^0.0.8"
flake8-annotations-coverage = "^0.0.6"
flake8-expression-complexity = "^0.0.11"
flake8-printf-formatting = "^1.1.2"
flake8-multiline-containers = "^0.0.19"
flake8-absolute-import = "^1.0.0.2"
flake8-simplify = "^0.21.0"
types-pytz = "^2024.1.0.20240203"
flake8-noqa = "^1.4.0"
pytest = "^8.1.1"
pytest-sugar = "^1.0.0"
pytest-cov = "^5.0.0"
pytest-factoryboy = "^2.7.0"
pytest-recording = "^0.13.1"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.6"
factory-boy = "^3.3.0"
async-factory-boy = "^1.0.1"
pytest-clarity = "^1.0.1"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
log_cli = "true"
log_cli_level = "INFO"
testpaths = [
    "src/my_apps",
]
pythonpath = [
    ".",
]
norecursedirs = ".mypy_cache .pytest_cache migrations"
filterwarnings = "ignore::DeprecationWarning"
asyncio_mode = "auto"